move getRasterTimeseriesAttributes and getVrtTimeseriesAttributes to raster_layer.py
