+++ src/map_click_handler.py
@@ -226,8 +226,8 @@ class TSClickHandler(MapClickHandler):
             self.ui.lb_msg_bar.setText(message)
             return
 
-        date_values, self.time_series_data = (
-            getRasterTimeseriesAttributes(layer, point=point, time_series_data=self.time_series_data))
+        date_values, self.time_series_data = raster_layer_utils.getRasterTimeseriesAttributes(
+            layer, point=point, time_series_data=self.time_series_data)
 
         if date_values.size == 0:
             return
@@ -251,73 +251,3 @@ class TSClickHandler(MapClickHandler):
         self.clearReferenceFeatureHighlight()
         self.plot_ts.plotTs(ref_values=self.ref_values)
 
-
-def getRasterTimeseriesAttributes(layer, point, time_series_data):
-    """
-    Get the timeseries values of the clicked point from the GMTSAR grd files.
-    The grd files should be in the same directory as the layer (typically velocity) file.
-    """
-    file_path = layer.source()
-    directory = os.path.dirname(file_path)
-
-    raster_file_paths, band_names = gmtsar_layer_utils.getGmtsarGrdInfo(directory)
-    dataset = raster_layer_utils.createVrtFromFiles(raster_file_paths=raster_file_paths,
-                                                    band_names=band_names, out_file="")
-
-    if not dataset:
-        return np.array([]), time_series_data
-
-    date_value_list, time_series_data = getVrtTimeseriesAttributes(dataset, point, time_series_data)
-    return date_value_list, time_series_data
-
-
-def getVrtTimeseriesAttributes(vrt_dataset, point, time_series_data, memory_limit=500):
-    """
-    Get the timeseries values of the clicked point from a vrt file that consists of time series data.
-    The vrt file should have description for each band in the format 'DYYYYMMDD'.
-    :param vrt_dataset: VRT dataset
-    :param point: QgsPointXY
-    :param time_series_data: numpy array
-    :param memory_limit: int in Mb
-    """
-
-    transform = vrt_dataset.GetGeoTransform()
-    inv_transform = gdal.InvGeoTransform(transform)
-
-    x, y = point.x(), point.y()
-    px, py = gdal.ApplyGeoTransform(inv_transform, x, y)
-    px, py = int(px), int(py)
-
-    band = vrt_dataset.GetRasterBand(1)
-    x_size = band.XSize
-    y_size = band.YSize
-    if not (0 <= px < x_size and 0 <= py < y_size):
-        return np.array([]), time_series_data
-
-    num_bands = vrt_dataset.RasterCount
-    data_type_size = gdal.GetDataTypeSize(band.DataType) // 8  # Size in bytes
-    expected_size = x_size * y_size * num_bands * data_type_size
-
-    if expected_size > memory_limit*1024*1024:
-        pixel_values = vrt_dataset.ReadAsArray(px, py, 1, 1)
-        if pixel_values is None:
-            return np.array([]), time_series_data
-        pixel_values = pixel_values[:, 0, 0]
-
-    else:  # read full data at once
-        if time_series_data is None:
-            time_series_data = vrt_dataset.ReadAsArray()
-        pixel_values = time_series_data[:, py, px]
-
-    if pixel_values is None:
-        return np.array([]), time_series_data
-
-    date_value_list = []
-    date_objs = [datetime.strptime(vrt_dataset.GetRasterBand(i).GetDescription()[1:], '%Y%m%d') for i in
-                 range(1, vrt_dataset.RasterCount + 1)]
-
-    for date_obj, pixel_value in zip(date_objs, pixel_values):
-        if not np.isnan(pixel_value):
-            date_value_list.append((date_obj, pixel_value))
-
-    return np.array(date_value_list, dtype=object), time_series_data
